fix: order change to ip address

+++ app/Http/Controllers/API/OrderAPIController.php
@@ -214,30 +214,32 @@ class OrderAPIController extends Controller
         $input = $request->all();
         $amount = 0;
         try {
-            $order = $this->orderRepository->create(
-                $request->only('user_id', 'order_status_id', 'tax', 'delivery_address_id', 'delivery_fee', 'hint')
-            );
-            Log::info($input['products']);
-            foreach ($input['products'] as $productOrder) {
-                $productOrder['order_id'] = $order->id;
-                $amount += $productOrder['price'] * $productOrder['quantity'];
-                $this->productOrderRepository->create($productOrder);
+            if ($request->input('delivery_address_id')) {
+                $order = $this->orderRepository->create(
+                    $request->only('user_id', 'order_status_id', 'tax', 'delivery_address_id', 'delivery_fee', 'hint')
+                );
+                Log::info($input['products']);
+                foreach ($input['products'] as $productOrder) {
+                    $productOrder['order_id'] = $order->id;
+                    $amount += $productOrder['price'] * $productOrder['quantity'];
+                    $this->productOrderRepository->create($productOrder);
+                }
+                $amount += $order->delivery_fee;
+                $amountWithTax = $amount + ($amount * $order->tax / 100);
+                $payment = $this->paymentRepository->create([
+                    "user_id" => $input['user_id'],
+                    "description" => trans("lang.payment_order_waiting"),
+                    "price" => $amountWithTax,
+                    "status" => 'Waiting for Client',
+                    "method" => $input['payment']['method'],
+                ]);
+
+                $this->orderRepository->update(['payment_id' => $payment->id], $order->id);
+
+                $this->cartRepository->deleteWhere(['user_id' => $order->user_id]);
+
+                Notification::send($order->productOrders[0]->product->market->users, new NewOrder($order));
             }
-            $amount += $order->delivery_fee;
-            $amountWithTax = $amount + ($amount * $order->tax / 100);
-            $payment = $this->paymentRepository->create([
-                "user_id" => $input['user_id'],
-                "description" => trans("lang.payment_order_waiting"),
-                "price" => $amountWithTax,
-                "status" => 'Waiting for Client',
-                "method" => $input['payment']['method'],
-            ]);
-
-            $this->orderRepository->update(['payment_id' => $payment->id], $order->id);
-
-            $this->cartRepository->deleteWhere(['user_id' => $order->user_id]);
-
-            Notification::send($order->productOrders[0]->product->market->users, new NewOrder($order));
 
         } catch (ValidatorException $e) {
             return $this->sendError($e->getMessage());
@@ -311,7 +313,7 @@ class OrderAPIController extends Controller
             }
 
         } catch (ValidatorException $e) {
-            return $this->sendError($e->getMessage());
+            return $this->sendError('Exception: '.$e->getMessage());
         }
 
         return $this->sendResponse($order->toArray(), __('lang.saved_successfully', ['operator' => __('lang.order')]));
